requirements.txt
streamlit==1.38.0
openai==1.51.2
python-dotenv==1.0.1
httpx==0.27.2